+++ changelog.md
@@ -1,7 +1,11 @@
 # Item Piles Changelog 
 
 ## Version 1.1.3
+- Adjusted display one-type item piles to also take into account dynamic attributes (gold piles!)
 - Fixed prototype tokens not being updated when editing an item pile through its sheet  
+- Fixed item piles with both "Is Container" and "Override single item token scale" enabled acting strange - item piles will now prioritize the container images over "Display Single Item Image" when "Is Container" is enabled
+- Added warning to point out the above
+- Adjusted Item Pile UI to be editable even when not enabled
 
 ## Version 1.1.2
 - Fixed dropping items onto piles not working when it had an interaction distance of infinite